Optimize requests in !rank
