+++ cogs/userrank.py
@@ -13,15 +13,14 @@ import libs.database as database
 from libs.embedmaker import officialEmbed
 from libs.utils import sanitize_check
 from libs.utils import api_fetch
+from libs.thm_api import get_sub_status, get_user_data
 
 
 ####################
 # Config variables #
 ####################
 
-c_api_rank = config.get_config("url")["api"]["user"]
 c_api_token = config.get_config("url")["api"]["token"]
-c_url_userprofile = config.get_config("url")["user_profile"]
 
 #####################
 # Strings variables #
@@ -39,52 +38,6 @@ def get_moto():
 
     return s_quotes[random.randint(0, len(s_quotes) - 1)]
 
-# Getting infos.
-def get_avatar(username):
-    """Fetches the user's avatar."""
-
-    response = requests.get(
-        c_api_rank.format(username))
-    data = response.text
-    data = json.loads(data)
-    return data['avatar']
-
-def get_points(username):
-    """Fetches the user's points."""
-
-    response = requests.get(
-        c_api_rank.format(username))
-    data = response.text
-    data = json.loads(data)
-    return data['points']
-
-def get_rank(username):
-    """Fetches the user's rank."""
-
-    response = requests.get(
-        c_api_rank.format(username))
-    data = response.text
-    data = json.loads(data)
-    return data['userRank']
-
-def get_sub_status(username):
-    """Fetches the user's sub status."""
-
-    url = c_url_userprofile.format(username)
-    check = "No!"
-    try:
-        response = requests.get(url)
-    except HTTPError as http_err:
-        print(f'HTTP error occurred: {http_err}')
-    except Exception as err:
-        print(f'Other error occurred: {err}')
-    else:
-        if "<span>Subscribed</span>" in response.text:
-            check = "Yes!"
-        else:
-            check = "No!"
-    return check
-
 
 ############
 # COG Body #
@@ -98,22 +51,24 @@ class Userrank(commands.Cog, name="Rank Commands"):
         """Sends the message containing the user's rank."""
 
         try:
-            if get_rank(user) != 0:
+            data = get_user_data(user)
+
+            if data['userRank'] != 0:
                 quip = get_moto()
                 quip = "*{}*".format(quip)
 
-                userImg = get_avatar(user)
-                Points = get_points(user)
-                rank = get_rank(user)
+                user_img = data['avatar']
+                points = data['points']
+                rank = data['userRank']
 
                 response = officialEmbed("Rank", quip, 0x148f77)
 
-                response.set_thumbnail(url=userImg)
+                response.set_thumbnail(url=user_img)
                 response.add_field(
                     name="Username:", value=user, inline=True)
                 response.add_field(name="Rank:", value=rank, inline=True)
                 response.add_field(
-                    name="Points:", value=Points, inline=True)
+                    name="Points:", value=points, inline=True)
 
                 sub = get_sub_status(user)
 
@@ -122,8 +77,8 @@ class Userrank(commands.Cog, name="Rank Commands"):
             else:
                 response = officialEmbed("Rank", s_userrank["user_not_found"], color=0xdc143c)
 
-                userImg = config.get_config("info")["icon"]
-                response.set_thumbnail(url=userImg)
+                user_img = config.get_config("info")["icon"]
+                response.set_thumbnail(url=user_img)
 
 
             await ctx.send(embed=response)
@@ -174,7 +129,6 @@ class Userrank(commands.Cog, name="Rank Commands"):
         if sanitize_check(user) == False:
             await ctx.send(config.get_string("not_sanitized"))
             return
-        
         if is_id:
             await self.rank_from_id(ctx, user)
         else:
